+++ cookbooks/wt_devicedataupdater/recipes/default.rb
@@ -1,41 +1,33 @@
 #
 # Cookbook Name:: wt_devicedataupdater
 # Recipe:: default
-# Author:: Tim Smith
 #
-# Copyright 2012, Webtrends
+# Copyright 2012, Webtrends Inc.
 #
 # All rights reserved - Do Not Redistribute
 #
 # This recipe installs the needed components to full setup/configure the Device Data Updater executeable
 #
 
-if ENV["deploy_build"] == "true" then
-  log "The deploy_build value is true so un-deploying first"
-  include_recipe "wt_devicedataupdater::uninstall"
+require 'win32/service'
+
+if ENV['deploy_build'] == 'true' then
+  log 'The deploy_build value is true so un-deploying first'
+  include_recipe 'wt_devicedataupdater::uninstall'
 else
-  log "The deploy_build value is not set or is false so we will only update the configuration"
+  log 'The deploy_build value is not set or is false so we will only update the configuration'
 end
 
 # get parameters
 download_url = node['wt_devicedataupdater']['download_url']
-master_host = node['wt_masterdb']['master_host']
 
 # destinations
-install_dir = File.join(node['wt_common']['install_dir_windows'], node['wt_devicedataupdater']['install_dir'].gsub(/[\\\/]+/,"\\"))
-log_dir = File.join(node['wt_common']['install_dir_windows'], node['wt_devicedataupdater']['log_dir'].gsub(/[\\\/]+/,"\\"))
+install_dir = File.join(node['wt_common']['install_dir_windows'], node['wt_devicedataupdater']['install_dir']).gsub(/[\\\/]+/,"\\")
+log_dir     = File.join(node['wt_common']['install_dir_windows'], node['wt_devicedataupdater']['log_dir']).gsub(/[\\\/]+/,"\\")
 
 auth_data = data_bag_item('authorization', node.chef_environment)
 svcuser = auth_data['wt_common']['system_user']
 svcpass = auth_data['wt_common']['system_pass']
-
-# determine root drive of install_dir - ENG390500
-if (install_dir =~ /^(\w:)\\.*$/)
-  install_dir_drive = $1
-else
-  raise Chef::Exceptions::AttributeNotFound,
-    "could not determine install_dir_drive, please verify value of install_dir: #{install_dir}"
-end
 
 # create the install directory
 directory install_dir do
@@ -43,11 +35,10 @@ directory install_dir do
   action :create
 end
 
-# set permissions for the service user to have read access to the install drive - ENG390500
-wt_base_icacls install_dir_drive do
-  action :grant
+wt_base_icacls install_dir do
   user svcuser
   perm :read
+  action :grant
 end
 
 # create the log directory
@@ -56,53 +47,37 @@ directory log_dir do
   action :create
 end
 
-# allow the service account to modify files in the log directory
 wt_base_icacls log_dir do
-  action :grant
   user svcuser
   perm :modify
+  action :grant
 end
 
 if deploy_mode?
 
-  ENV['deploy_build'] = 'false'
-  # log("Source URL: #{build_url}") { level :info}
-  #this_zipfile = get_build node['wt_devicedataupdater']['zip_file']
+  # Platform Scheduler Agent must be running
+  ruby_block 'WebtrendsAgent_status' do
+    block do
+      raise 'WebtrendsAgent is not running' unless Win32::Service.status('WebtrendsAgent').current_state == 'running' 
+    end
+  end
 
- # unzip the install package
+  ENV['deploy_build'] = 'false'
+  log "Source URL: #{download_url}"
+
+  # unzip the install package
   windows_zipfile install_dir do
       source download_url
       action :unzip
   end
 
-
-  # unzip the install package
-  #windows_zipfile install_dir do
-   # source "#{this_zipfile}"
-    #action :unzip
-  #end
-
-template "#{install_dir}\\DDU.exe.config" do
-    source "DeviceDataUpdater.erb"
-    variables(
-    )
-end
-
-powershell "install device data updater" do
-    environment({'install_dir' => install_dir, 'service_binary' => node['wt_devicedataupdater']['service_binary']})
-        code <<-EOH
-        $binary_path = $env:install_dir + "\\" + $env:service_binary
-        &$binary_path --install
-        EOH
-end
-
-template "#{install_dir}\\log4net.config" do
-    source "log4net.erb"
-    variables(
-        :logdir => log_dir
-    )
-end
+  # full path to DDU.exe
+  ddu = File.join(install_dir, node['wt_devicedataupdater']['service_binary']).gsub(/[\\\/]+/,"\\")
+  execute node['wt_devicedataupdater']['service_binary'] do
+    command "#{ddu} -install"
+  end
 
   share_wrs
+
 end