+++ cookbooks/wt_streamingapi/recipes/undeploy.rb
@@ -29,12 +29,12 @@ service "streamingapi" do
   ignore_failure true
 end
 
-directory "#{log_dir}" do
+directory log_dir do
   recursive true
   action :delete
 end
 
-directory "#{install_dir}" do
+directory install_dir do
   recursive true
   action :delete
 end