+++ cookbooks/wt_streamingcollection/attributes/default.rb
@@ -12,4 +12,5 @@ default['wt_streamingcollection']['group']        = "webtrends"
 default['wt_streamingcollection']['download_url'] = ""
 default['wt_streamingcollection']['port']         = 8080
 default['wt_streamingcollection']['java_opts']    = "-Xms1024m -Djava.net.preferIPv4Stack=true"
-default['wt_streamingcollection']['jmx_port']     = 9999
+default['wt_streamingcollection']['jmx_port']     = 10000
+default['wt_streamingcollection']['kafka_connect_timeout_ms'] = 100