+++ cookbooks/wt_sync/metadata.rb
@@ -3,6 +3,6 @@ maintainer_email "kendrick.martin@webtrends.com"
 license          "All rights reserved"
 description      "Installs/Configures Webtrends Sync Service"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "1.0.4"
+version          "1.0.5"
 depends          "wt_base"
 depends          "powershell"