+++ cookbooks/wt_xd/recipes/schema.rb
@@ -18,7 +18,7 @@ end
 # drop in templates
 %w[hbasetable.py schema.py].each do |file|
   cookbook_file "#{install_dir}/#{file}" do
-    source "${file}"
+    source "#{file}"
     owner 'hadoop'
     group 'hadoop'
     mode 00550